Refactor to 1 class per file
